fiddle with signal comp demo
Performance is still abysmally bad. There are times when just rendering
two graphs will drop the frames to 15-20 fps. It's confusing because
it's inconsistent and goes away sometimes if graphics are released
and recreated.

Thoughts:
- precreate graphics and hide them?
- experiment with custom graphics

Note:
I've found that sometimes graphics require the beginFill() operation.
It completely baffles me and I believe it's related to how the
Graphics object interacts with the RenderContext. I believe I'm misusing
graphics entirely and it might be worth experimenting with creating
my own Lines object for rendering tons of lines or similar.